Update Page_template.md
job #1 - update template to remove http and https from the video link.  Having this in there messes up the embedding when the site uses the different form.  By not specifying the http type in the link, the website will use whatever it is currently running under and the link will work properly in either case.